dynamic shapes
I coded and designed this project using the p5.js library. The movement
of the string of cubes is controlled by the user’s mouse movements. In
addition, as the user moves their mouse within the canvas, the
background color fades from blue, to grey, to tan seamlessly. The only
element that is independently mobile is the smallest black cube at the
end of the string.